back/Reports/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
python-multipart==0.0.6
